+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/Column.java
@@ -150,12 +150,12 @@ public sealed interface Column<T> extends Attribute<T>, ColumnConditions<T> perm
 	 * Creates a new {@link Column}, associated with the given entityType.
 	 * @param entityType the entityType owning this column
 	 * @param name the column name
-	 * @param valueClass the class representing the column value type
+	 * @param type the class representing the column value type
 	 * @param <T> the column type
 	 * @return a new {@link Column}
 	 */
-	static <T> Column<T> column(EntityType entityType, String name, Class<T> valueClass) {
-		return new DefaultColumn<>(name, valueClass, entityType);
+	static <T> Column<T> column(EntityType entityType, String name, Class<T> type) {
+		return new DefaultColumn<>(name, type, entityType);
 	}
 
 	/**