+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRenderers.java
@@ -69,7 +69,7 @@ public class EntityTableCellRenderers implements FilterTableCellRenderer.Factory
 	 */
 	protected <T> FilterTableCellRenderer.Builder<Entity, Attribute<?>, T> builder(AttributeDefinition<T> attributeDefinition) {
 		return configure(requireNonNull(attributeDefinition), FilterTableCellRenderer.<Entity, Attribute<?>>builder()
-						.type(attributeDefinition.attribute().type().valueClass()));
+						.type(attributeDefinition.attribute().type().get()));
 	}
 
 	static <T> FilterTableCellRenderer.Builder<Entity, Attribute<?>, T> configure(AttributeDefinition<T> attributeDefinition,